# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Knouz Arabic E-Commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, popular, price_asc, price_desc]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: featured
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /products/{id}/reviews:
    get:
      operationId: listProductReviews
      tags: [reviews]
      summary: Get reviews for a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
...
components:
  schemas:
...
    Category:
      type: object
      required: [id, nameAr, slug]
      properties:
        id:
          type: string
        nameAr:
          type: string
        slug:
          type: string
        image:
          type: ["string", "null"]
        productCount:
          type: ["integer", "null"]
...
    Product:
      type: object
      required: [id, nameAr, descriptionAr, price, stock, images, categoryId]
      properties:
        id:
          type: string
        nameAr:
          type: string
        descriptionAr:
          type: string
        price:
          type: number
        comparePrice:
          type: ["number", "null"]
        stock:
          type: integer
        images:
          type: array
          items:
            type: string
        categoryId:
          type: string
        category:
          $ref: "#/components/schemas/Category"
        featured:
          type: boolean
        avgRating:
          type: ["number", "null"]
        reviewCount:
          type: ["integer", "null"]
        createdAt:
          type: string
    ProductInput:
      type: object
      required: [nameAr, descriptionAr, price, stock, images, categoryId]
      properties:
        nameAr:
          type: string
        descriptionAr:
          type: string
        price:
          type: number
        comparePrice:
          type: ["number", "null"]
        stock:
          type: integer
        images:
          type: array
          items:
            type: string
        categoryId:
          type: string
        featured:
          type: boolean
    ProductUpdate:
      type: object
      properties:
        nameAr:
          type: string
        descriptionAr:
          type: string
        price:
          type: number
        comparePrice:
          type: ["number", "null"]
        stock:
          type: integer
        images:
          type: array
          items:
            type: string
        categoryId:
          type: string
        featured:
          type: boolean
    ProductList:
      type: object
      required: [products, total, page, totalPages]
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
    Review:
      type: object
      required: [id, productId, rating]
      properties:
        id:
          type: string
        userId:
          type: string
        userName:
          type: ["string", "null"]
        productId:
          type: string
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
    ReviewInput:
      type: object
      required: [rating]
      properties:
        rating:
          type: integer
        comment:
          type: string
        userId:
          type: string
        userName:
          type: string
...