# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Knouz Arabic E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from './category';

export interface Product {
  id: string;
  nameAr: string;
  descriptionAr: string;
  price: number;
  /** @nullable */
  comparePrice?: number | null;
  stock: number;
  images: string[];
  categoryId: string;
  category?: Category;
  featured?: boolean;
  /** @nullable */
  avgRating?: number | null;
  /** @nullable */
  reviewCount?: number | null;
  createdAt?: string;
}