# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Knouz Arabic E-Commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  nameAr: string;
  descriptionAr: string;
  price: number;
  /** @nullable */
  comparePrice?: number | null;
  stock: number;
  images: string[];
  categoryId: string;
  featured?: boolean;
}